+++ package.json
@@ -10,9 +10,12 @@
   },
   "dependencies": {
     "@emotion/cache": "^11.14.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
     "@fullcalendar/core": "^6.1.17",
     "@fullcalendar/daygrid": "^6.1.17",
     "@fullcalendar/react": "^6.1.17",
+    "@mui/material": "^7.0.2",
     "@mui/material-nextjs": "^7.0.2",
     "@tanstack/react-query": "^5.74.7",
     "next": "15.3.1",
