+++ src/app/auth/login/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <div>로그인 페이지</div>;
+}
