+++ src/app/auth/page.tsx
