+++ src/app/layout.tsx
@@ -1,0 +1,26 @@
+import type { Metadata } from "next";
+import "./globals.css";
+import MainHeader from "@/components/MainHeader";
+import { AppRouterCacheProvider } from "@mui/material-nextjs/v15-appRouter";
+
+export const metadata: Metadata = {
+  title: "Looper",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body>
+        <AppRouterCacheProvider>
+          <MainHeader />
+          <main>{children}</main>
+        </AppRouterCacheProvider>
+      </body>
+    </html>
+  );
+}
