+++ src/components/MainHeader.module.css
@@ -1,0 +1,29 @@
+.header {
+  width: calc(100% - 40px);
+  max-width: 100%;
+  height: 60px;
+  background-color: var(--color-background);
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0 20px;
+  border-bottom: 1px solid #e0e0e0;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.header a {
+  text-decoration: none;
+  color: var(--color-primary);
+  font-family: var(--logo-font);
+  font-weight: 750;
+  font-size: 30;
+}
+
+.logo {
+  display: flex;
+  justify-content: space-between;
+  width: 135px;
+  height: 50px;
+  line-height: 50px;
+  text-decoration: none;
+}
